requirements.txt
streamlit
python-dotenv
requests
PyPDF2